% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_featurecodes.R
\docType{data}
\name{featurecodes}
\alias{featurecodes}
\title{Geonames Feature Codes table for Feature classes & types}
\format{{A data.frame with 7 variables
\describe{
\itemize{
\item {\code{feature_full:}}{ The full feature code}
\item {\code{feature_class:}}{ The class of feature:
\item A = country, state, region,
\item H = stream, lake,
\item L = parks, area,
\item P = city, village,
\item R = road, railroad,
\item S = spot, building, farm,
\item T = mountain, hill, rock,
\item U = undersea,
\item V = forest, heath}
\item {\code{feature_code:}}{ Specific feature codes, character}
\item {\code{feature_name:}}{ Type of feature, character}
\item {\code{feature_detail:}}{ Details of a feature}
\item {\code{feature_clean:}}{ Processed feature name field with brackets and plurals removed, character}
\item{\code{multi:}}{ Processed field indicating the presence of a plural in feature name, logical}
}
}
}}
\source{
\href{http://download.geonames.org/export/dump/featureCodes_en.txt}{geonames featureCodes_en.txt}
}
\usage{
data("featurecodes")
}
\description{
The english geonames featurecodes table \href{http://download.geonames.org/export/dump/featureCodes_en.txt}{geonames featurecode}.
}
\examples{
featurecodes
}
\references{
Please cite as: GeoNames. http://geonames.org/.
}
\keyword{datasets}
