% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_KE.R
\docType{data}
\name{KE}
\alias{KE}
\title{Kenya}
\format{{
A data frame with 29598 observations on the following 21 variables.
\describe{
\item{\code{geonameid}}{a character vector}
\item{\code{name}}{a character vector}
\item{\code{asciiname}}{a character vector}
\item{\code{alternatenames}}{a character vector}
\item{\code{latitude}}{a character vector}
\item{\code{longitude}}{a character vector}
\item{\code{feature_full}}{a character vector}
\item{\code{feature_class}}{a character vector}
\item{\code{feature_code}}{a character vector}
\item{\code{iso}}{a character vector}
\item{\code{cc2}}{a character vector}
\item{\code{admin1_code}}{a character vector}
\item{\code{admin2_code}}{a character vector}
\item{\code{admin3_code}}{a character vector}
\item{\code{admin4_code}}{a character vector}
\item{\code{population}}{a character vector}
\item{\code{elevation}}{a character vector}
\item{\code{dem}}{a numeric vector}
\item{\code{timezone}}{a character vector}
\item{\code{modification_date}}{a Date}
\item{\code{download_date}}{a Date}
}
}}
\usage{
data("KE")
}
\description{
Kenya geonames dataset
}
\keyword{datasets}
