% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\docType{data}
\name{places}
\alias{places}
\title{Places R6 class}
\description{
The Places generator creates a new \code{Places}-object, which is the class containing
all the app logic. The class is based on the \link[R6:R6Class]{R6} OO-system and
is thus reference-based with methods and data attached to each object, in
contrast to the more well known S3 and S4 systems.
}
\section{Initialization}{

A new 'Places'-object is \link[places:places-init]{initialized} using the \code{new()} method on the generator:

\tabular{l}{
\code{dirs <- places$new()}
}
}

\section{Methods}{


\tabular{ll}{
\link[places:places-add]{$add} \tab Add or Update paths to the object \cr
\link[places:places-get]{$get} \tab Return stored paths by name \cr
\link[places:places-names]{$names} \tab Return the names of the stored paths \cr
\link[places:places-rm]{$rm} \tab Remove stored paths \cr
\link[places:places-filepath]{$filepath} \tab Construct filepaths using stored paths as the root \cr
}
}

\keyword{datasets}
