% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{.add}
\alias{.add}
\alias{places-add}
\title{Add/Update stored path}
\usage{
.add(self, private, ...)
}
\arguments{
\item{self}{places self object}

\item{private}{places private object}

\item{\dots}{named elements to store}
}
\description{
Add or Update a stored path in the places object
}
\examples{
 paths <- places$new()
 paths
 paths$add(back = '..',this='.')
 paths
 paths$names()
}
\seealso{
\link[places:places]{places}
@importFrom stats setNames
}
