% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places_countryinfo.R
\name{places_countryinfo}
\alias{places_countryinfo}
\title{Import the Country Info file}
\usage{
places_countryinfo()
}
\value{
data.frame
}
\description{
The geonames country info text file \url{http://download.geonames.org/export/dump/countryInfo.txt } contains the country names and a range of codes along with information linked to the capital city.
}
\details{
Column names are converted to lowercase, spaces are replaced with underscores and brackets removed.
}
\examples{
{places_countryinfo()}
}
