% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places_download.R
\name{places_download}
\alias{places_download}
\title{Download the latest Country data}
\usage{
places_download(country = NULL, code = NULL)
}
\arguments{
\item{country}{A country name (use \code{\link{places_lookup}}). For all countries use "allcountries" (note > 300 MB file).}

\item{code}{A two letter iso country code (use \code{\link{places_lookup}})}
}
\value{
a data.frame
}
\description{
Read in and process geonames data dump files into a data.frame.
}
\details{
Column names are converted to lower, spaces replaced with underscores and brackets removed. download_date added to table with the users system date. \code{places_import} cannot handle three letter country codes will throw an error.
}
\examples{
\dontrun{AI <- places_download(code = "AI") \%>\% places_import()
kenya <- places_download(country = "kenya") \%>\% places_import()
all <- places_download(country = "allcountries") \%>\% places_import()
}
}
