% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{.filepath}
\alias{.filepath}
\alias{places-filepath}
\title{Create file.paths}
\usage{
.filepath(self, private, path, ..., mustWork = TRUE)
}
\arguments{
\item{self}{places self object}

\item{private}{places private object}

\item{path}{name of stored path}

\item{\dots}{additional character vectors}

\item{mustWork}{logical that is passed to \link[base:normalizePath]{normalizePath}, Default: TRUE}
}
\description{
Use stored paths as the root of a file.path call
}
\examples{
 dir <- places$new()
 root <- basename(dir$root)
 dir$add(back = '..',this='.')
 dir$filepath('back',root)

 #set mustWork to FALSE to create new directories or files
 dir$filepath('back','somefile.R',mustWork = FALSE)
}
\seealso{
\link[places:places]{places}
}
