% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get}
\alias{.get}
\alias{places-get}
\title{Get stored paths}
\usage{
.get(self, private, ...)
}
\arguments{
\item{self}{places self object}

\item{private}{places private object}

\item{\dots}{names of stored paths}
}
\value{
single name returns a character, multiple names or empty returns a list
}
\description{
Get a path or a list of paths stored in places object
}
\examples{
 dir <- places$new()
 dir
 dir$add(back = '..',this='.')
 dir$get()
 dir$get('back')
 dir$get(c('back','this'))
}
\seealso{
\link[places:places]{places}
}
