% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{places_gps}
\alias{places_gps}
\title{GPS coordinates.}
\format{
A data frame with 309 rows and 7 variables:
\describe{
  \item{user_id}{unique identifier for each participant}
  \item{time_local}{datetime of GPS coordinates}
  \item{lat}{latitude}
  \item{lng}{longitude}
  \item{ema}{survey report id}
  \item{Response.Time}{datetime of survey report}
  \item{tz_olson_id}{timezone label}
  ...
}
}
\usage{
places_gps
}
\description{
A dataset containing the GPS coordinates and other attributes for 1 hypothetical person.
}
\keyword{datasets}
