% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places_import.R
\name{places_import}
\alias{places_import}
\title{Import downloaded geonames files}
\usage{
places_import(file)
}
\arguments{
\item{file}{A character vector containing a file name with a .zip extension returned by \code{\link{places_download}}}
}
\value{
tibble
}
\description{
Imports files downloaded with \code{\link{places_download}}
}
\examples{
\dontrun{AI <- places_download(code = "AI") \%>\% places_import()
AI <- places_download(country = "Anguilla") \%>\% places_import()
res <- places_download(code = AI)
AI <- places_import(res)}
}
