% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.init}
\alias{.init}
\alias{places-init}
\title{Initialize places object}
\usage{
.init(self, private, root = here::here())
}
\arguments{
\item{self}{places self object}

\item{private}{places private object}

\item{root}{root path of the object, Default: here::here()}
}
\value{
print method for the places object
}
\description{
Initialize new places object with root, Default: here::here()
}
\details{
root path is tested via \link[base:normalizePath]{normalizePath} that it exists.
}
\examples{
 places$new()
 places$new(here::here('..'))
}
\seealso{
\link[places:places]{places}, \link[here:here]{here}
}
