% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places_lookup.R
\name{places_lookup}
\alias{places_lookup}
\title{Helper to look up Country Codes to Import}
\usage{
places_lookup(name = NULL, type = "country")
}
\arguments{
\item{name}{A country, region or subregion name.}

\item{type}{Either "country", "region", "subregion" or "intermediate"}
}
\value{
A vector containing a country name or set of two letter country codes for a region or subregion.
}
\description{
A helper function to navigate two letter country codes. The geonames dump file \code{\link{places_table}} uses two letter country codes (iso2c). The \code{\link{places_table}} includes optional country and region names. If you prefer to look up country codes then use this helper. The function wraps the \code{\link[countrycode]{countrycode}} lookup function for matching single country names or vectors to two letter country codes. Country and region names are not case sensitive.
}
\details{
The region, subregion and intermediate searches are mainly intended for use with the allcountries table to group countries by United Nations regions. The function wraps the \href{https://github.com/vincentarelbundock/countrycode}{countrycode} package on CRAN function \code{\link[countrycode]{countrycode}}
\itemize{ UN Regions
\item Africa
\item Americas
\item Asia
\item Europe
\item Oceania
}
\itemize{ UN Subregion
\item Australia and New Zealand
\item Central Asia
\item Eastern Asia
\item Eastern Europe
\item Latin America and the Caribbean
\item Melanesia
\item Micronesia
\item Northern Africa
\item Northern America
\item Northern Europe
\item Polynesia
\item South-eastern Asia
\item Southern Asia
\item Southern Europe
\item Sub-Saharan Africa
\item Western Asia
\item Western Europe
}
\itemize{UN Intermediate Region
\item Caribbean
\item Central America
\item Channel Islands
\item Eastern Africa
\item Middle Africa
\item South America
\item Southern Africa
\item Western Africa
}
}
\examples{
\dontrun{one <- places_lookup("UK", type = "country"),
two <- places_lookup(c("US", "UK"), type = "country"),
two_names <- places_lookup(c("Kenya", "Tanzania"))
upper <- places_lookup("Africa", type = "region")
africa <- places_lookup("africa", type = "region"),
central_asia <- places_lookup("Central Asia", type = "subregion"),
caribbean <- places_lookup("Caribbean", type = "intermediate")
}
}
