% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.names}
\alias{.names}
\alias{places-names}
\title{Get names of stored paths}
\usage{
.names(self, private)
}
\arguments{
\item{self}{places self object}

\item{private}{places private object}
}
\description{
Get names of the stored paths in the places object
}
\examples{
 paths <- places$new()
 paths
 paths$add(back = '..',this='.')
 paths$names()
}
\seealso{
\link[places:places]{places}
}
