% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{.rm}
\alias{.rm}
\alias{places-rm}
\title{Remove a stored path}
\usage{
.rm(self, private, ...)
}
\arguments{
\item{self}{places self object}

\item{private}{places private object}

\item{\dots}{name(s) of stored path(s)}
}
\description{
Remove a stored path in the places object
}
\examples{
 paths <- places$new()
 paths
 paths$add(back = '..',this='.')
 paths
 paths$names()
 paths$rm('back')
 paths$names()
}
\seealso{
\link[places:places]{places}
}
