% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.setwd}
\alias{.setwd}
\alias{places-setwd}
\title{Set working directory using a stored path}
\usage{
.setwd(self, private, place)
}
\arguments{
\item{self}{places self object}

\item{private}{places private object}

\item{place}{name of stored path}
}
\description{
Set working directory using a stored path
}
\examples{
 paths <- places$new()
 paths
 paths$add(back = '..',this='.')
 paths$setwd('back')
 paths$setwd('this')
}
\seealso{
\link[places:places]{places}
}
