% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{read_path}
\alias{read_path}
\title{handle file import from file path}
\usage{
read_path(path)
}
\arguments{
\item{path}{geonames file path}
}
\description{
Import raw file and return df with regular cols where path specified. Note that may need to adjust for windows file path cases.
}
