% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{read_places}
\alias{read_places}
\title{Read in file}
\usage{
read_places(x)
}
\arguments{
\item{x}{Input from places_import}
}
\description{
Take the input from places_import, download the file, name the columns and return a data frame.
}
