% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Combined Regions Table}
\format{{
\describe{
\item{\code{un_global_code}}{numeric}
\item{\code{un_global_name}}{character}
\item{\code{un_region_code}}{numeric}
\item{\code{un_region_name}}{character}
\item{\code{un_sub_region_code}}{numeric}
\item{\code{un_sub_region_name}}{character}
\item{\code{un_intermediate_region_code}}{numeric}
\item{\code{un_intermediate_region_name}}{character}
\item{\code{un_country_or_area}}{character}
\item{\code{un_m49_code}}{character}
\item{\code{iso3}}{character}
\item{\code{un_least_developed_countries_ldc}}{numeric}
\item{\code{un_land_locked_developing_countries_lldc}}{numeric}
\item{\code{un_small_island_developing_states_sids}}{numeric}
\item{\code{un_developed_or_developing_countries}}{character}
\item{\code{iso}}{character}
\item{\code{iso_numeric}}{numeric}
\item{\code{fips}}{character}
\item{\code{country}}{character}
\item{\code{geonameid}}{numeric}
\item{\code{wb_region}}{World Bank region, character}
\item{\code{wb_income}}{World Bank income, character}
}
}}
\usage{
regions
}
\value{
data.frame
}
\description{
United Nations Statistical Division and World Bank Development Indicator regions
}
\examples{
regions
}
\keyword{datasets}
