% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_unique_places.R
\docType{data}
\name{unique_places}
\alias{unique_places}
\title{Unique Place Names}
\format{{
A data frame with 7,230,431 observations of three variables.
\describe{
\item{\code{asciiname}}{a place name, character vector}
\item{\code{n}}{number of occurrences across all countries, numeric}
\item{\code{ngram}}{number of ngrams making up a place name, character vector}
}
}}
\usage{
data("unique_places")
}
\description{
A data frame of unique asciiname entries for place names in
the allcountries geonames file. It is used for text mining.
}
\keyword{datasets}
