% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wdicountry.R
\docType{data}
\name{wdicountry}
\alias{wdicountry}
\title{World Development Indicators Regions}
\format{{
A data frame with 304 observations on the following 3 variables.
\describe{
\item{\code{iso3}}{a character vector}
\item{\code{wb_region}}{a character vector}
\item{\code{wb_income}}{a character vector}
}
}}
\usage{
wdicountry
}
\description{
From the WDI package
}
\examples{
wdicountry
}
\keyword{datasets}
