% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\docType{package}
\name{places}
\alias{places}
\alias{places-package}
\title{\code{places} package}
\description{
Access geonames daily files
}
\details{
See the README on
\href{https://github.com/poldham/places#readme}{GitHub}
}
