% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places_import.R
\name{places_import}
\alias{places_import}
\title{Import Geonames Data Files}
\usage{
places_import(country = NULL, code = NULL, path = NULL)
}
\arguments{
\item{country}{A country name (use \code{\link{places_lookup}}). For all countries use "all" (note > 300 MB file).}

\item{code}{A two letter iso country code (use \code{\link{places_lookup}})}

\item{path}{For manually downloaded files as either zip or txt.}
}
\value{
a data.frame
}
\description{
Read in and process geonames data dump files into a data.frame.
}
\details{
Column names are converted to lower, spaces replaced with underscores and brackets removed. download_date added to table with the users system date.
}
\examples{
\dontrun{cc <- places_import(code = "AI")}
}
