% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.names}
\alias{.names}
\alias{places-names}
\title{Get names of stored paths}
\usage{
.names(self, private)
}
\description{
Get names of the stored paths in the places object
}
\examples{
 dir <- places$new()
 dir
 dir$add(back = '..',this='.')
 dir$names()
}
\seealso{
\link[places:places]{places}
}
