% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{.rm}
\alias{.rm}
\alias{places-rm}
\title{Remove a stored path}
\usage{
.rm(self, private, obj)
}
\arguments{
\item{self}{places self object}

\item{private}{places private object}

\item{\dots}{named elements to remove from storage}
}
\description{
Remove a stored path in the places object
}
\examples{
 dir <- places$new()
 dir
 dir$add(back = '..',this='.')
 dir
 dir$names()
 dir$rm('back')
 dir$names()
}
\seealso{
\link[places:places]{places}
}
